Move download tests
